+++ biobox_cli/command/verify.py
@@ -2,47 +2,89 @@
 biobox verify - Verify that a Docker image matches the given specification type
 
 Usage:
-    biobox verify <biobox_type> <image> [--task=TASK]
+    biobox verify <biobox_type> <image> [--task=TASK] [--verbose] [--log=FILE]
 
 Options:
   -h, --help             Show this screen.
-  -t TASK --task=TASK    Specify which biobox task to test. [default: default]
+  -t TASK, --task=TASK   Specify which biobox task to test. [default: default]
+  -v, --verbose          Show the status of each biobox verification test.
+  -l FILE, --log=FILE    Log test results to file.
 
 Available Biobox types:
 
   short_read_assembler  Assemble short reads into contigs
 """
 
-import biobox_cli.util as util
-import sys, os, os.path, tempfile
+import biobox_cli.util.misc        as util
+import biobox_cli.util.error       as error
+import biobox_cli.util.functional  as fn
+import biobox_cli.behave_interface as behave
+import biobox_cli.container        as ctn
 
-def verification_file(biobox):
-    from pkg_resources import resource_filename
-    file_ = biobox + '.feature'
-    return os.path.abspath(resource_filename(__name__, os.path.join('..', 'verification', file_)))
+import string
+import sys
 
-def tmp_feature_dir():
-    return os.path.abspath(os.path.join(os.getcwd(), 'biobox_verify'))
+from fn    import F, _
+from fn.op import flip
+
+def format_scenario_name(name):
+    return fn.thread([
+        string.replace(name, "Should ", ""),
+        lambda x: x[0].upper() + x[1:],
+        F(flip(string.split), '--'),
+        fn.first,
+        string.strip])
+
+def format_scenario_status(status):
+    formats = {
+            'passed'  : 'PASS',
+            'failed'  : 'FAIL',
+            'not run' : 'NOT RUN'}
+    return formats[status]
 
 def run(argv):
-    opts   = util.parse_docopt(__doc__, argv, False)
-    biobox = opts['<biobox_type>']
-    image  = opts['<image>']
-    task   = opts['--task']
+    opts    = util.parse_docopt(__doc__, argv, False)
+    biobox  = opts['<biobox_type>']
+    image   = opts['<image>']
+    task    = opts['--task']
+    verbose = opts['--verbose']
+    log     = opts['--log']
 
-    from behave.__main__ import main as behave_main
-    _, tmp_file = tempfile.mkstemp()
-    cmd = "{file} --define IMAGE={image} --define TASK={task} --define TMPDIR={tmp_dir} --outfile {tmp_file} --no-summary --stop"
-    args = {'file':     verification_file(biobox),
-            'tmp_dir':  tmp_feature_dir(),
-            'image':    image,
-            'tmp_file': tmp_file,
-            'task':     task}
+    if not behave.features_available(biobox):
+        error.err_exit("unknown_command",
+                {"command_type" : "biobox type", "command" : biobox})
 
-    behave_main(cmd.format(**args))
+    ctn.exit_if_no_image_available(image)
 
-    with open(tmp_file, 'r') as f:
-        output = f.read()
+    if verbose:
+        results = behave.run(biobox, image, task, False)
+    else:
+        results = behave.run(biobox, image, task)
 
-    if "Assertion Failed" in output:
-        util.err_exit('failed_verification', {'image': image, 'biobox': biobox.replace('_', ' ')})
+    if verbose:
+        if log:
+            sys.stdout = open(log, "w+")
+        statuses = fn.thread([
+            behave.get_scenarios_and_statuses(results),
+            F(map, lambda (x, y): (format_scenario_name(x), format_scenario_status(y)))])
+        longest_name = fn.thread([
+            statuses,
+            F(map, fn.first),
+            F(map, len),
+            max])
+        output = fn.thread([
+            statuses,
+            F(map, lambda (x, y): (string.ljust(x, longest_name, ' '), y)),
+            F(map, F(flip(string.join), "   ")),
+            fn.unique,
+            F(flip(string.join), "\n")])
+        print output
+    elif behave.is_failed(results):
+        if log:
+            sys.stderr = open(log, "w+")
+        msg = fn.thread([
+            behave.get_failing_scenarios(results),
+            F(map, behave.scenario_name),
+            F(flip(string.join), "\n")])
+
+        error.err_exit('failed_verification', {'image': image, 'error': msg, 'biobox' : biobox})