+++ biobox_cli/verification/steps/cli.py
@@ -25,6 +25,9 @@ def assert_file_not_empty(file_):
     with open(file_, 'r') as f:
         assert_not_empty(f.read().strip())
 
+def remove_warnings(string_):
+    return re.sub("^WARNING:.+\n", "", string_)
+
 
 
 @given(u'I create the directory "{directory}"')
@@ -40,6 +43,12 @@ def step_impl(context, file_):
 def step_impl(context):
     for row in context.table.rows:
         shutil.copy(get_data_file_path(row['source']),
+                get_env_path(context, row['dest']))\
+
+@given(u'I copy the example data directories')
+def step_impl(context):
+    for row in context.table.rows:
+        shutil.copytree(get_data_file_path(row['source']),
                 get_env_path(context, row['dest']))
 
 @when(u'I run the command')
@@ -53,6 +62,12 @@ def step_impl(context):
 def step_impl(context, stream):
     output = get_stream(context, stream)
     nt.assert_equal(output, "",
+            "The {} should be empty but contains:\n\n{}".format(stream, output))
+
+@then(u'excluding warnings the {stream} should be empty')
+def step_impl(context, stream):
+    output = get_stream(context, stream)
+    nt.assert_equal(remove_warnings(output), "",
             "The {} should be empty but contains:\n\n{}".format(stream, output))
 
 @then(u'the exit code should be {code}')
@@ -71,11 +86,21 @@ def step_impl(context, stream):
     output = get_stream(context, stream)
     nt.assert_equal(context.text, output)
 
+@then(u'excluding warnings the {stream} should equal')
+def step_impl(context, stream):
+    output = get_stream(context, stream)
+    nt.assert_equal(context.text, remove_warnings(output))
+
 @then(u'the {stream} should match /{regexp}/')
 def step_impl(context, stream, regexp):
     output = get_stream(context, stream)
     nt.assert_not_equal(None, re.match(regexp, output),
       "Regular expression {} not found in:\n'{}'".format(regexp, output))
+
+@then(u'the directory "{}" should not exist')
+def step_impl(context, dir_):
+    nt.assert_false(os.path.isdir(dir_),
+            "The directory \"{}\" should not exist.".format(dir_))
 
 @then(u'the file "{}" should exist')
 def step_impl(context, file_):